Corrections for using water quality modules with group macros.

+++ crhmcode/src/modules/waterquality/ClassGrow_crops_annually.h
@@ -103,7 +103,7 @@ public:
 	double* fertday_down{ NULL };
 	double* resday_up{ NULL };
 	double* resday_down{ NULL };
-	long* fertperiod{ NULL };
+	double* fertperiod{ NULL };   // must match WQ_Soil_NPCcycling declaration (PRL)
 	long* litterperiod{ NULL };
 
 	// declared observation variables
